+++ packages/internal.ts
@@ -9,6 +9,18 @@
 import { InitializeGame } from '../src/core/initialize';
 import { ProcessGameConfig } from '../src/core/game';
 import { CreateGameReducer } from '../src/core/reducer';
+import { getFilterPlayerView } from '../src/master/filter-player-view';
 import { Async, Sync } from '../src/server/db/base';
+import { Transport } from '../src/client/transport/transport';
+import { createMatch } from '../src/server/util';
 
-export { Async, Sync, ProcessGameConfig, InitializeGame, CreateGameReducer };
+export {
+  Async,
+  Sync,
+  Transport,
+  ProcessGameConfig,
+  InitializeGame,
+  CreateGameReducer,
+  createMatch,
+  getFilterPlayerView,
+};