+++ src/server/util.ts
@@ -33,7 +33,7 @@ export const createMetadata = ({
 };
 
 /**
- * Creates matchID, initial state and metadata for a new match.
+ * Creates initial state and metadata for a new match.
  * If the provided `setupData` doesn’t pass the game’s validation,
  * an error object is returned instead.
  */
@@ -60,4 +60,26 @@ export const createMatch = ({
   const initialState = InitializeGame({ game, numPlayers, setupData });
 
   return { metadata, initialState };
+};
+
+/**
+ * Given players, returns the count of players.
+ */
+export const getNumPlayers = (players: Server.MatchData['players']): number =>
+  Object.keys(players).length;
+
+/**
+ * Given players, tries to find the ID of the first player that can be joined.
+ * Returns `undefined` if there’s no available ID.
+ */
+export const getFirstAvailablePlayerID = (
+  players: Server.MatchData['players']
+): string | undefined => {
+  const numPlayers = getNumPlayers(players);
+  // Try to get the first index available
+  for (let i = 0; i < numPlayers; i++) {
+    if (typeof players[i].name === 'undefined' || players[i].name === null) {
+      return String(i);
+    }
+  }
 };